Step 3
Step 3 : Now open app -> java -> package -> MainActivity.java and add the below code.

In this step we used the functions that linked via the button click. These functions are defined in other class and are used here. Each function return value that define no of rows updated, using that we defined whether operation is successful or not. Also user need to define valid data to perform operation empty fields will not be entertained and return error .