Create Step 4
Step 4: In this step create a java class myDbAdapter. java.

In this we define the functions that are used to perform the operations insert, update and delete operations in SQLite. Further this class create another class that will extend the SQLiteOpenHelper. Each function carry equivalent methods that perform operations.

Important Note – According to naming convention it is suggested to define primary key starting with underscore example: _id.